restructure resident page
