+++ src/app/resident-page/resident-info-page/resident-info-page.component.html
@@ -1,0 +1,1 @@
+<h1>Hello</h1>
