+++ src/app/resident-page/resident-info-page/resident-info-page.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+@Component({
+  selector: 'app-resident-info-page',
+  templateUrl: './resident-info-page.component.html',
+  styleUrls: ['./resident-info-page.component.scss'],
+})
+export class ResidentInfoPage {}
