+++ src/app/resident-page/resident-list-page/resident-list-page.component.ts
@@ -1,9 +1,9 @@
-import { ResidentInfo } from './../shared/list/list';
+import { ResidentService } from './../../service/resident.service';
+import { FireStoreService } from './../../service/firestore.service';
+import { ResidentInfo } from './../../shared/list/list';
 import { Subscription } from 'rxjs';
-import { ResidentService } from './../service/resident.service';
-import { FireStoreService } from './../service/firestore.service';
 import { Component, OnDestroy } from '@angular/core';
-import { emit } from 'process';
+
 @Component({
   selector: 'app-resident-list-page',
   templateUrl: './resident-list-page.component.html',
@@ -31,6 +31,7 @@ export class ResidentListPage implements OnDestroy {
   onResident(email: string) {
     console.log(email);
   }
+
   onCheckIn(email: string) {
     this.fireStoreService.checkInResident(email);
   }
