+++ src/app/resident-page/resident-page.component.ts
@@ -1,0 +1,6 @@
+import { Component } from '@angular/core';
+@Component({
+  selector: 'app-resident-page',
+  template: '<router-outlet></router-outlet>',
+})
+export class ResidentPage {}
